+++ CHANGELOG.md
@@ -61,6 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - #562: Fixed #553. Handle visualization of empty graphs.
 
+- #561: Fixed #555. Method `Pattern.draw` computes Pauli flow from pattern.
+
 ### Changed
 
 - #452: Use `uv` for dependency management